cd C++

+++ C++/math/primefactortable.hpp
@@ -8,9 +8,9 @@ struct p_fact {
     vectors::vi spf;
     p_fact(const int n): spf(n + 1){
         std::iota(spf.begin(), spf.end(), 0);
-        for(int i = 2; i * i <= n; ++i) {
+        for(int64_t i = 2; i * i <= n; ++i) {
             if(spf[i]==i) {
-                for(int j = i * i; j <= n; j += i) {
+                for(int64_t j = i * i; j <= n; j += i) {
                     if(spf[j]==j) spf[j]=i;
                 }
             }
