+++ C++/math/primetable.hpp
@@ -6,9 +6,9 @@ struct p_table {
     vectors::vb SoE;
     p_table(const int n): SoE(n+1,1){
         SoE[0]=SoE[1]=0;
-        for(int i = 2; i <= n; ++i) {
+        for(int64_t i = 2; i <= n; ++i) {
             if(!SoE[i]) continue;
-            for(int j = i * i; j <= n; j += i) SoE[j] = 0;
+            for(int64_t j = i * i; j <= n; j += i) SoE[j] = 0;
         }
     }
     vectors::vi get() {
